Struct taskpipe::TaskPipe

pub struct TaskPipe<T: 'static + Send> {
    // some fields omitted
}

A taskpipe continuation

This struct is not meant to be used directly. Instead, it the return type of other functions within this library

Methods

impl<T: Send> TaskPipe<T>

fn input<F: 'static + FnOnce(Sender<T>) + Send>(self, task: F) -> TaskPipe<T>

Insert additional items into the pipeline

Items from the previous task will be passed through to the next task. Items produced by the given function will also be received by downstream tasks.

use std::sync::mpsc::Sender;

taskpipe::input(|tx: Sender<char>| {
    for c in "abcdefghijklmnopqrstuvwxyz".chars() { tx.send(c).unwrap(); }
}).input(|tx: Sender<char>| {
    for c in "0123456789".chars() { tx.send(c).unwrap(); }
});

fn pipe<S: 'static + Send, F: 'static + FnOnce(Receiver<T>, Sender<S>) + Send>(self, task: F) -> TaskPipe<S>

Transform or process pipeline items

Received upstream task items, process them, and pass them along to downstream tasks.

use std::sync::mpsc::{Receiver,Sender};

taskpipe::input(|tx: Sender<char>| {
    for c in "abcdefghijklmnopqrstuvwxyz".chars() { tx.send(c).unwrap(); }
}).pipe(|rx: Receiver<char>, tx: Sender<bool>| {
    for c in rx.iter() {
        tx.send(c.is_uppercase()).unwrap();
    }
});

fn end<'a, R: Send + 'a, F: 'static + FnOnce(Receiver<T>) -> R + Send>(self, task: F) -> JoinGuard<'a, R>

Finalize the pipeline, consuming any items.

This method dispatches and returns a Thread JoinGuard while the closure receives any items that have made their way through the entire pipeline.

Example

use std::sync::mpsc::{Receiver,Sender};

let guard = taskpipe::input(|tx: Sender<char>| {
    for c in "abcdefghijklmnopqrstuvwxyz".chars() {
        tx.send(c).unwrap();
    }
}).end(|rx: Receiver<char>| {
    for c in rx.iter() {
        print!("{} ", c);
    }
});

Example with result

use std::sync::mpsc::{Receiver,Sender};

let result = taskpipe::input(|tx: Sender<char>| {
    for c in "abcdefghijklmnopqrstuvwxyz".chars() { tx.send(c).unwrap(); }
}).end(|rx: Receiver<char>| {
    for c in rx.iter() {
        print!("{} ", c);
    }
}).join();